4.21*: bloglist expansion, step9

+++ controllers/blogs.js
@@ -54,11 +54,23 @@ blogsRouter.post('/', async (request, response) => {
 
 // trycatch is here for 4.13 Blog list expansions, step1, usually handled by express-async-errors
 blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
-  } catch (error) {
-    next(error);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (user._id.toString() === blog.user.toString()) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } catch (error) {
+      next(error);
+    }
+  } else {
+    return response.status(401).json({ error: 'operation unauthorized' });
   }
 });
